Backend/FastAPI/requirements.txt
fastapi[standard]
python-jose
passlib
bcrypt
pymongo
argon2_cffi
dotenv